Stagefright: Allow setting high-framerates in CameraSource
CameraSource (legacy Camera) treats framerate set via
MediaRecorder.setVideoFrameRate as _both_ camera-preview-fps
and video-encoder-fps. Trying to set high fps for encoder
will inadvertently fail trying to set the same value for
camera-preview.
Read the custom camera parameter to detect if this is a
high-speed request, selectively set the fps only for
video, and default to 30fps for preview fps.

Change-Id: If66211dd81b2a08d4df4c6f23e87304e9e7013f4
Signed-off-by: xyyx <xyyx@mail.ru>
Signed-off-by: Anirudh Gupta <anirudhgupta109@gmail.com>